+++ src/blockchain/script.cpp
@@ -1481,7 +1481,7 @@ public:
         // (~200 bytes per cache entry times 50,000 entries)
         // Since there are a maximum of 20,000 signature operations per block
         // 50,000 is a reasonable default.
-        int64_t nMaxCacheSize = GetArg("-maxsigcachesize", 50000);
+        int64_t nMaxCacheSize = GetArg("-maxsigcachesize", (int64_t) 50000);
         if (nMaxCacheSize <= 0) return;
 
         LOCK(cs_sigcache);
