+++ src/wallet/wallet.cpp
@@ -3916,8 +3916,9 @@ bool CWallet::NewKeyPool()
         if (IsLocked())
             return false;
 
-        int64_t nKeys = max(GetArg("-keypool", chainParams.DEFAULT_KEYPOOL),
-                            (int64_t)0);
+        int64_t nKeys = max(GetArg("-keypool",
+                                   (int64_t) chainParams.DEFAULT_KEYPOOL),
+                            (int64_t) 0);
         for (int i = 0; i < nKeys; i++)
         {
             int64_t nIndex = i+1;
@@ -3949,8 +3950,9 @@ bool CWallet::TopUpKeyPool(unsigned int nSize)
         }
         else
         {
-            nTargetSize = max(GetArg("-keypool", chainParams.DEFAULT_KEYPOOL),
-                              (int64_t)0);
+            nTargetSize = max(GetArg("-keypool",
+                                     (int64_t) chainParams.DEFAULT_KEYPOOL),
+                              (int64_t) 0);
         }
 
         while (setKeyPool.size() < (nTargetSize + 1))
